refactor(123): Split RideMenu into smart/view components and fix PR comments
